+++ src/app/camps/camps.page.ts
@@ -67,28 +67,28 @@ function initialState(): CampsState {
 }
 
 @Component({
-    selector: 'app-camps',
-    templateUrl: 'camps.page.html',
-    styleUrls: ['camps.page.scss'],
-    changeDetection: ChangeDetectionStrategy.OnPush,
-    imports: [
-        CommonModule,
-        RouterModule,
-        ScrollingModule,
-        MapModalComponent,
-        IonHeader,
-        IonToolbar,
-        IonTitle,
-        IonButtons,
-        IonContent,
-        IonText,
-        IonBadge,
-        CampComponent,
-        SearchComponent,
-        CategoryComponent,
-        AlphabeticalScrollBarComponent,
-        SortComponent,
-    ]
+  selector: 'app-camps',
+  templateUrl: 'camps.page.html',
+  styleUrls: ['camps.page.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  imports: [
+    CommonModule,
+    RouterModule,
+    ScrollingModule,
+    MapModalComponent,
+    IonHeader,
+    IonToolbar,
+    IonTitle,
+    IonButtons,
+    IonContent,
+    IonText,
+    IonBadge,
+    CampComponent,
+    SearchComponent,
+    CategoryComponent,
+    AlphabeticalScrollBarComponent,
+    SortComponent,
+  ]
 })
 export class CampsPage {
   public db = inject(DbService);